Move rider functions in :drive RideRepository to :ride RideRequestRepository
